small updates

+++ include/Logger.hpp
@@ -174,9 +174,10 @@ namespace ssrlcv{
     /**
      * Used to return a high resolution timestamp for dense resolution timing of algorithms
      * this is not nessesary needed for a final logger and coult be substituted for a less
-     * precise time measurement 
+     * precise time measurement
+     * @return timestamp the number of milliseconds sense January 1st 1970
      */
-    double getTime();
+    unsigned long getTime();
 
   }; // end Logger class
 }
